moves validators and middlewares to their respective folders
