+++ src/tests/conversion/stream.endpoint.test.ts
@@ -9,7 +9,7 @@ import Upload from "../../database/models/upload";
 import { SSEvents } from "../../controllers/conversion";
 import { VALID_CONVERT_TARGETS, INVALID_CONVERT_TARGET } from "../utils";
 import { AmazonS3 } from "../../helpers/awsS3";
-import { ConversionMiddleware } from "../../middlewares/conversion";
+import { ConversionHelper } from "../../helpers/conversion";
 
 config();
 chai.use(chaiHttp);
@@ -49,8 +49,8 @@ describe("GET /stream/:id", () => {
                 });
             });
 
-        ConversionMiddleware.CALL_INTERVAL = 10;
-        ConversionMiddleware.INCREASE_BY = 50;
+        ConversionHelper.CALL_INTERVAL = 10;
+        ConversionHelper.INCREASE_BY = 50;
         const { fileId } = await createUpload.get({ plain: true });
 
         const spySSESend = sinon.spy(SSEvents.prototype, "send");
